+++ BlissMixerExt/Settings.pm
@@ -19,6 +19,7 @@ use Slim::Utils::Strings qw(string);
 use Slim::Utils::Versions;
 
 my $prefs = preferences('plugin.blissmixerext');
+my $upstreamPrefs = preferences('plugin.blissmixer');
 my $serverprefs = preferences('server');
 
 sub name {
@@ -30,7 +31,8 @@ sub page {
 }
 
 sub prefs {
-    return ($prefs, 'learned_blend', 'playcount_influence', 'triplets_backup_path');
+    return ($prefs, 'learned_blend', 'playcount_influence',
+        'lastfm_track_guidance_percent', 'triplets_backup_path');
 }
 
 sub beforeRender {
@@ -50,6 +52,12 @@ sub beforeRender {
     $paramRef->{database_exists} = -e File::Spec->catfile($dbDir, 'bliss.db') ? 1 : 0;
     $paramRef->{matrix_exists} = -e File::Spec->catfile($dbDir, 'learned_matrix.json') ? 1 : 0;
     $paramRef->{statistics_enabled} = main::STATISTICS ? 1 : 0;
+    $paramRef->{lastmix_available} = Slim::Utils::PluginManager->isEnabled(
+        'Plugins::LastMix::Plugin'
+    ) ? 1 : 0;
+    $paramRef->{upstream_lastfm_enabled} =
+        $upstreamPrefs->get('use_adaptive_weights')
+        && $upstreamPrefs->get('use_lastfm_weighting') ? 1 : 0;
     $paramRef->{no_learner_binary} = !Slim::Utils::Misc::findbin('bliss-learner');
     $paramRef->{learning_start_text} = string('BLISSMIXEREXT_LEARNING_START_TIME');
     $paramRef->{learning_duration_text} = string('BLISSMIXEREXT_LEARNING_DURATION');
@@ -65,6 +73,18 @@ sub beforeRender {
 
 sub handler {
     my ($class, $client, $paramRef) = @_;
+    for my $setting (
+        ['pref_learned_blend', 0, 100],
+        ['pref_playcount_influence', -100, 100],
+        ['pref_lastfm_track_guidance_percent', 0, 100],
+    ) {
+        my ($name, $minimum, $maximum) = @$setting;
+        next unless defined $paramRef->{$name};
+        my $value = int($paramRef->{$name});
+        $value = $minimum if $value < $minimum;
+        $value = $maximum if $value > $maximum;
+        $paramRef->{$name} = $value;
+    }
     return $class->SUPER::handler($client, $paramRef);
 }